monitor-py/requirements.txt
Kafka-python
